2015-11.py: Improve increment_password interface, usage
The `test_new_password` test was hitting a case where the
list of characters returned from increment_password wasn’t
joined into a string before being used where a string was
expected (a regex match).

This was a deficiency of the interface of increment_password,
where the burden should be on the function to prepare the
current password for incrementing and then rejoin it,
not on the caller to use `list()` and then `.join()`.