2019 10 Python: Fix relative_distance, atan2 use (again)
I worked through some examples with atan2 to understand what varying its paramaters does (swapping y and x, and inverting either or both).

Turns out you can adjust from which ray (quarter hour) the angle starts, and in which direction it proceeds.

The negation of atan2 gives us -pi to +pi so we don't have to reverse the sort (which might not be so obviously important).