2019 12 Python: Use plain tuples instead of NamedTuple for performance
This affects Position and Velocity, which were mostly just wrappers around Tuple[int, int, int] with named fields.

This has been the most significant 'single' change. With the previous loop-unrolling it's reduced the running time by about 40%.

Moon is still a NamedTuple because the clarity it brings outways the performance disadvantage.