2019 13 Python: Vastly improve part two performance
Previous, naive, approach was to search the state dictionary for the ball and paddle position each time the bot made a move. This was incredibly slow, O(n * m) where n is the number of positions on the screen and m is roughly the number of frames until game over.

Now, we track changes to the game state (hence the rename in a previou commit) and update saved ball and paddle positions accordingly.